Simplify inheritance layers (#639)
* remove LanguageModel

* remove TransformerDecoder

* remove TransforemrDecoderLayer

* lint update

* remove MultiheadAttention -> ModelParallelMultiheadAttention

* updated the test_hf_compatibility.py

* update test documentation

---------
