+++ metaseq/models/__init__.py
@@ -16,7 +16,6 @@ from .distributed_model import DistributedModel
 from .base_decoder import BaseDecoder
 from .base_model import (
     BaseModel,
-    LanguageModel,
 )
 
 
@@ -32,7 +31,6 @@ __all__ = [
     "BaseModel",
     "DistributedModel",
     "BaseDecoder",
-    "LanguageModel",
     "register_model",
 ]
 
@@ -91,11 +89,11 @@ def register_model(name, dataclass=None):
     For example::
 
         @register_model('lstm')
-        class LSTM(LanguageModel):
+        class LSTM(BaseModel):
             (...)
 
     .. note:: All models must implement the :class:`BaseModel` interface.
-        Typically you will extend :class:`LanguageModel` for
+        Typically you will extend :class:`BaseModel` for
         language modeling tasks.
 
     Args:
