+++ metaseq/modules/__init__.py
@@ -8,14 +8,12 @@ from .dropout import Dropout
 from .activation_functions import ActivationFn, gelu
 from .layer_norm import Fp32LayerNorm, LayerNorm
 from .learned_positional_embedding import LearnedPositionalEmbedding
-from .multihead_attention import MultiheadAttention
-from .multihead_attention_mp import ModelParallelMultiheadAttention
+from .multihead_attention import ModelParallelMultiheadAttention
 from .positional_embedding import PositionalEmbedding
 from .sinusoidal_positional_embedding import SinusoidalPositionalEmbedding
 from .linear import Linear
 from .feedforward_network import FeedForwardNetwork
 from .transformer_decoder_layer import (
-    TransformerDecoderLayer,
     ModelParallelTransformerDecoderLayer,
 )
 from .sequence_parallel_transformer_layer import SequeuceParallelTransformerBlock
@@ -27,13 +25,11 @@ __all__ = [
     "gelu",
     "LayerNorm",
     "LearnedPositionalEmbedding",
-    "MultiheadAttention",
     "ModelParallelMultiheadAttention",
     "PositionalEmbedding",
     "SinusoidalPositionalEmbedding",
     "Linear",
     "FeedForwardNetwork",
-    "TransformerDecoderLayer",
     "ModelParallelTransformerDecoderLayer",
     "SequeuceParallelTransformerBlock",
 ]
