+++ metaseq/modules/transformer_decoder_layer.py
@@ -12,12 +12,10 @@ from torch import Tensor
 from metaseq import utils
 from metaseq.modules import (
     ActivationFn,
-    MultiheadAttention,
     ModelParallelMultiheadAttention,
     Dropout,
     FeedForwardNetwork,
     LayerNorm,
-    Linear,
 )
 from metaseq.modules.fused_bias_gelu import (
     has_fused_bias_gelu,
@@ -39,11 +37,10 @@ def _weight_init(weight):
     return nn.init.kaiming_uniform_(weight, a=math.sqrt(5))
 
 
-class TransformerDecoderLayer(nn.Module):
-    """Pre-norm Decoder layer block.
-
+class ModelParallelTransformerDecoderLayer(nn.Module):
+    """Decoder layer block.
     Note that we have found model training to require pre-norm to remain stable.
-
+    See "Megatron-LM: https://arxiv.org/pdf/1909.08053.pdf" for more details.
     Args:
         args (argparse.Namespace): parsed command-line arguments
     """
@@ -95,7 +92,6 @@ class TransformerDecoderLayer(nn.Module):
             "truncate_init": getattr(args, "truncate_init", False),
         }
 
-        # Note: ModelParallelTransformerDecoderLayer overrides build_fc1.
         self.fc1 = self.build_fc1(
             self.embed_dim,
             ffn_dim,
@@ -110,7 +106,6 @@ class TransformerDecoderLayer(nn.Module):
             **fc1_kwargs,
         )
 
-        # Note: ModelParallelTransformerDecoderLayer overrides build_fc2.
         self.fc2 = self.build_fc2(
             ffn_dim,
             self.embed_dim,
@@ -128,149 +123,6 @@ class TransformerDecoderLayer(nn.Module):
         self.final_layer_norm.to(device).to(dtype)
         self.args = args
 
-    # Refer to model_parallel's transformer layer for why fc1 and fc2 are separate methods.
-    def build_fc1(
-        self,
-        input_dim,
-        output_dim,
-        initialize_params_on_gpu=False,
-        disable_bias=False,
-        **unused_args
-    ):
-        return Linear(
-            input_dim,
-            output_dim,
-            initialize_params_on_gpu=initialize_params_on_gpu,
-            bias=not disable_bias,
-            dtype=utils.get_model_init_dtype(self.args),
-        )
-
-    def build_fc2(
-        self,
-        input_dim,
-        output_dim,
-        initialize_params_on_gpu=False,
-        disable_bias=False,
-        **unused_args
-    ):
-        return Linear(
-            input_dim,
-            output_dim,
-            initialize_params_on_gpu=initialize_params_on_gpu,
-            dtype=utils.get_model_init_dtype(self.args),
-        )
-
-    def build_self_attention(
-        self,
-        embed_dim,
-        args,
-        add_bias_kv=False,
-        add_zero_attn=False,
-    ):
-        return MultiheadAttention(
-            embed_dim,
-            args.decoder_attention_heads,
-            dropout=args.attention_dropout,
-            add_bias_kv=add_bias_kv,
-            add_zero_attn=add_zero_attn,
-            self_attention=True,
-            initialize_params_on_gpu=getattr(
-                args, "tensor_parallel_init_model_on_gpu", False
-            ),
-            bias=not getattr(
-                args,
-                "disable_bias",
-                False,
-            ),
-            dtype=utils.get_model_init_dtype(args),
-        )
-
-    def forward_attention(
-        self,
-        query,
-        key,
-        value,
-        residual,
-        key_padding_mask=None,
-        incremental_state=None,
-        attn_mask=None,
-    ):
-        x, _ = self.self_attn(
-            query=query,
-            key=key,
-            value=value,
-            key_padding_mask=key_padding_mask,
-            incremental_state=incremental_state,
-            attn_mask=attn_mask,
-        )
-        x = self.dropout_module(x)
-        x = residual + x
-        return x
-
-    def forward(
-        self,
-        x,
-        incremental_state: Optional[Dict[str, Dict[str, Optional[Tensor]]]] = None,
-        self_attn_mask: Optional[torch.Tensor] = None,
-        self_attn_padding_mask: Optional[torch.Tensor] = None,
-        recompute_fc1: bool = False,
-    ):
-        """
-        Args:
-            x (Tensor): input to the layer of shape `(seq_len, batch, embed_dim)`
-
-        Returns:
-            encoded output of shape `(seq_len, batch, embed_dim)`
-        """
-        if getattr(self.args, "sequence_parallel", False):
-            from metaseq.modules import SequeuceParallelTransformerBlock
-
-            x = SequeuceParallelTransformerBlock.apply(
-                x,
-                self.self_attn.qkv_proj.weight,
-                self.self_attn.out_proj.weight,
-                self.fc1.weight,
-                self.fc2.weight,
-                self.self_attn.head_dim,
-                recompute_fc1,
-                self.activation_fn_name,
-            )
-            return x
-
-        residual = x
-        x = self.self_attn_layer_norm(x)
-        x = self.forward_attention(
-            query=x,
-            key=x,
-            value=x,
-            residual=residual,
-            key_padding_mask=self_attn_padding_mask,
-            incremental_state=incremental_state,
-            attn_mask=self_attn_mask,
-        )
-        residual = x
-        x = self.final_layer_norm(x)
-        x = FeedForwardNetwork(
-            x,
-            fc1=self.fc1,
-            activation_fn=self.activation_fn,
-            fc2=self.fc2,
-            dropout_module=self.dropout_module,
-        )
-        x = residual + x
-        return x
-
-    def make_generation_fast_(self, **kwargs):
-        pass
-
-
-class ModelParallelTransformerDecoderLayer(TransformerDecoderLayer):
-    """Decoder layer block.
-
-    See "Megatron-LM: https://arxiv.org/pdf/1909.08053.pdf" for more details.
-    """
-
-    # TODO[susanz]: unify method signatures with non-model-parallel version.
     def build_fc1(
         self,
         input_dim,
@@ -314,8 +166,6 @@ class ModelParallelTransformerDecoderLayer(TransformerDecoderLayer):
             bias=not disable_bias,
         )
 
-    # TODO[susanz]: unify method signatures with non-model-parallel version.
-    # Note: only fc2 includes the full_megatron_init_scalar arg, given the scaled_init_method_normal call.
     def build_fc2(
         self,
         input_dim,
@@ -415,3 +265,59 @@ class ModelParallelTransformerDecoderLayer(TransformerDecoderLayer):
         )
         x = x + residual
         return x
+
+    def forward(
+        self,
+        x,
+        incremental_state: Optional[Dict[str, Dict[str, Optional[Tensor]]]] = None,
+        self_attn_mask: Optional[torch.Tensor] = None,
+        self_attn_padding_mask: Optional[torch.Tensor] = None,
+        recompute_fc1: bool = False,
+    ):
+        """
+        Args:
+            x (Tensor): input to the layer of shape `(seq_len, batch, embed_dim)`
+
+        Returns:
+            encoded output of shape `(seq_len, batch, embed_dim)`
+        """
+        if getattr(self.args, "sequence_parallel", False):
+            from metaseq.modules import SequeuceParallelTransformerBlock
+
+            x = SequeuceParallelTransformerBlock.apply(
+                x,
+                self.self_attn.qkv_proj.weight,
+                self.self_attn.out_proj.weight,
+                self.fc1.weight,
+                self.fc2.weight,
+                self.self_attn.head_dim,
+                recompute_fc1,
+                self.activation_fn_name,
+            )
+            return x
+
+        residual = x
+        x = self.self_attn_layer_norm(x)
+        x = self.forward_attention(
+            query=x,
+            key=x,
+            value=x,
+            residual=residual,
+            key_padding_mask=self_attn_padding_mask,
+            incremental_state=incremental_state,
+            attn_mask=self_attn_mask,
+        )
+        residual = x
+        x = self.final_layer_norm(x)
+        x = FeedForwardNetwork(
+            x,
+            fc1=self.fc1,
+            activation_fn=self.activation_fn,
+            fc2=self.fc2,
+            dropout_module=self.dropout_module,
+        )
+        x = residual + x
+        return x
+
+    def make_generation_fast_(self, **kwargs):
+        pass
